requirements.txt
google-genai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
statsmodels>=0.14.0
matplotlib>=3.7.0
flask>=3.0.0
flask-cors>=4.0.0
